Fixed a bug when converting emp to bgeo, where if a channel with an unsupported type is encountered, a segmentation fault occurred.
Added support for Int64Type channels ( since the particle::id64 is of that type ).
Since Houdini does not yet support 64bit int attributes, we split it 64bit channels
values into a size 2 int attribute in Houdini, where the least and most significant bits
are stored in index 0 and 1 respectively.

We also now create a local variable mapping ($NAME) for all channels brought in as attributes
in Houdini/bgeos.